OnlySaveOnTerminate --> SaveConstantly
This way OnlySaveOnTerminate behavior when set to true is now the
default behavior. Since it's a boolean, this was the only way to do that
(without using a different variable type). Since I can't detect a
difference between the user setting it to False and it being set to
False by default.

Also I removed a comment about update checking, no plans to maintain
this project. If that changes could add an update system then.