Added DontProcessVideos
-Added DontProcessVideos boolean, which if set to true would result in
the recorder not turning the images in temp_images into a video. I
thought this might save on cpu usage a lot, thus making SaveConstantly a
more viable option but alas this did not seem to be the case. Saving the
images to disk seems to really hit the cpu, atleast a lot more than I
thought it was. Being able to stream the images to ffmpeg should save on
cpu usage notably... if somebody else wants to look into that. Kind of
just leaving this in the code since I guess it doesn't hurt.

-Also some minor cosmetic changes with some writelns